+++ changelog.md
@@ -1,4 +1,4 @@
-# Upcoming
+# 5.12.0.0
 * Added Scarce Copper
   * Makes copper ores drop 1\~2 copper instead of 2\~5 (about 58% less)
 * Tweaks